+++ tetris_game/src/lib.rs
@@ -1,5 +1,7 @@
 mod common;
 mod tetromino;
+mod playfield;
+mod states;
 mod tetris;
 
 pub use common::{Position, Color, Button, GamePad, GameUI};
